# Make high resolution scrolling more responsive and configurable