I was testing high resolution scrolling under Wayland and found that scrolling becomes acceptable if I multiply `yoffset` by `10` and awesome if I multiply it by `20`.

Since `wheel_scroll_multiplier` is not used under high resolution scrolling at all, I thought we could put it to good use here, so that responsiveness becomes configurable.

Under Wayland, my `viewport_y_ratio` is `2`. I added it to this patch because it was suggested by @Luflosi in #1112, if you want to keep it, https://github.com/kovidgoyal/kitty/commit/82f9aecacbc0f76ca9e7aec809fc8cf3fd73ff5a probably needs to be reverted.

With this formula, scrolling is acceptable with default settings and I can further improve it if I change `wheel_scroll_multiplier` to `10`.

If we remove `viewport_y_ratio` from the formula, the default value of `wheel_scroll_multiplier` (`5`) is too slow, and I would need to change it to at least `10`, but better to `20`.