@Luflosi 2018-11-09
At least on macOS a value of 1 makes kitty scroll as fast as all other apps.