@maximbaz 2018-11-09
Just to confirm, are you testing with https://github.com/kovidgoyal/kitty/commit/82f9aecacbc0f76ca9e7aec809fc8cf3fd73ff5a reverted? Because both that commit and this PR increase the speed on macOS.

In any case, I don't really care about the default value, I'm totally fine changing it in my local config as long as it's configurable 😉