@Luflosi 2018-11-09
That commit doesn't make a difference for me since `yscale` is 1 for me since I don't have a retina display. And I didn't suggest changing the default value for everyone, only where high precision scrolling is supported. I think that 5 lines per scroll event are pretty reasonable when scrolling with a mouse that doesn't support high precision scrolling.