@maximbaz 2018-11-09
Maybe it wasn't a good idea then to re-use `wheel_scroll_multiplier` option and we better have a different one for high precision scrolling...