@Luflosi 2018-11-09
I think the semantics of that option make perfect sense for low and high precision scrolling and we should not create a new option.