@maximbaz 2018-11-09
If the option will have different values based on whether the device is "high" or "low" precision, users would need to have a way to override _both_ these values, and this will be impossible I think...

In other words, I would like to override `wheel_scroll_multiplier` to `10` to get the speed as in sakura with touchpad, but I don't want to change the default `5` for when I use a mouse.