@kovidgoyal 2018-11-10
It would be interesting to see what VTE does on wayland. Why is it scrolling faster than kitty? Does it multiply the pixels by something and if so where does the value of that something come from? @egmontkob you know the answer?