@egmontkob 2018-11-10
GTK+ gives us a smooth scrolling delta which hides from us the X11 vs. Wayland differences, and is probably already scaled by some global GTK+ or GNOME setting (I'm not sure about it). I'm not aware of these details.

Then in VTE it's scaled according to ceil(height / 10.0) [here](https://gitlab.gnome.org/GNOME/vte/blob/0.54.2/src/vte.cc#L9872), which may not be a good approach as per [748012](https://bugzilla.gnome.org/show_bug.cgi?id=748012). See also [769696](https://bugzilla.gnome.org/show_bug.cgi?id=769696).