@ducis 2020-03-11
Hello, I am wondering, in the current master branch, when the last line is updated (or a new line is added to the bottom), does kitty redraw the whole window? like if there is an "A" in the second line from the bottom, then the first line from the bottom changed, does kitty go through the process of rasterizing the letter "A" again?  I assume if it does not need to (by holding a buffer of some previously rasterized parts of the terminal window), both high-prec scrolling would be easy and less computation would be needed?