@ducis 2020-03-11
So far as I know, a terminal program either is only able to change some lines from the bottom (normal terminal program, and tab completion) or takes control of the whole (virtual) terminal (like vim, tmux, or clear), it does not change lines from the top or in the middle. In the latter case you don't have scrolling anyway. So we can always assume a terminal program changes a certain number of lines from the bottom. Is it correct?