@kovidgoyal 2020-03-11
kitty only ever rasterizes a character once, that's part of the magic
behind its performance. And no nay terminal program
can make changes to any part of the screen at any time, for example by
simply changing the background color, or directly by moving the cursor
around the screen.