@ducis 2020-03-11
Thanks. Then where shall I start looking into the scrolling/rendering stuff? In particular, which variables in the source refer to the previously rasterized content?