@kovidgoyal 2020-03-11
there is no previously rasterized content. rendering happens on the GPU.
If you want to get previously rendered content you have to render into a
framebuffer. There is code that does that for other reasons in shaders.c
but you are going to need familiarity with OpenGL to make it work