@kovidgoyal 2020-03-11
On Tue, Mar 10, 2020 at 08:07:16PM -0700, ducis. wrote:
> So by "rasterizes a character once" you meant there are caches like many images with the size of individual characters in the graphics memory?

There is a sprite map stored on the GPU.

> 
> I guess by "does that for other reasons in shaders.c" you mean draw_cells_interleaved_premult().
> Is this function triggered in, say, a fresh install of kitty without customized configuration file or passed commandline options? If not, how should I enable it so that I can test if it affects the latency first?

It will be used when you have a transparent window and images displayed
in the window.