@kovidgoyal 2020-03-11
On Tue, Mar 10, 2020 at 08:20:13PM -0700, ducis. wrote:
> Does the framebuffer to render into live in the graphics memory or the main memory? 
> Usually for this kind of things it should live in the graphics memory and the application has pointer-like stuff with which the app can tell the GPU to do things like memcpy() purely in the graphics memory.
> But I am not sure if you meant we have to fetch the rendered results into the main memory by "you have to render into a framebuffer", which would be much worse performance-wise.

framebuffers live in GPU memory