@ducis 2020-03-11
Hi, I just tested with typometer.
The environment is compiz (0.8.*, not the Ubuntu one) 60Hz frame rate globally, 
both delays turned to 0, VSync enabled in compiz but disabled in kitty (so no tearing).
Transparency within kitty roughly doubles the latency from ~17ms to ~31ms,
**with or without icatted images.**
But transparency with compiz plugins (which can make any window translucent) results in latency ~26ms.
Would you think that the 17ms->31ms is caused by going through draw_cells_interleaved_premult()?
Am I missing anything?