@kovidgoyal 2020-03-11
Put a printf() in draw_cells_interleaved_premult() to check if it is going through that. IIRC without images it will not be used, see line 636 onwards of shaders.c for details. Either you need transparency + any image or non-transparent and negative z-index images.