@ducis 2020-03-11
I put in the printf() and tested again. Yes, draw..premult() is only triggered by both background_opacity and icat.
How transparency affects latency really depended on what windows were below it.
It seems that the more complex stuff are below a translucent window the more likely the computation will take more than one frame.
But in general transparency,
as well as draw_cells_interleaved_premult() when triggered by icat, do not force an additional frame.