@kovidgoyal 2020-03-11
You should be able to trigger it with just

kitty -o background_image=whatever.png

and no transparency (assuming you are running from master)